Add files via upload
The openai file has the code to run the abstract to code algorithm on given paper 10 times. Every time it generates a questions text file and a code text file which has all the compucell3d codes for cc3d, py, xml and steppable py files. The second file compares run 2 to 10 code outputs to the first run and rates it 1 to 5. 1 is lowest similarity and 5 is highest similarity. Also, it gives a brief description of the differences between the 2 codes.